CoercibleToAssetSelection (#13132)
## Summary & Motivation

`define_asset_job` has some fancy code for coercing a variety of
arguments into `AssetSelection`.

Factor this out so it can be used elsewhere.

## How I Tested These Changes
